backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
numpy==2.2.2
fastembed==0.4.2
google-api-python-client==2.179.0
google-auth==2.40.3
pydantic-settings==2.10.1
